Update
I spent some time styling the landing page to further the process of finishing this challenge along. I resized the hero image, the logo, wrangled the h1 and p elements so they match the design line for line. Next I have to connect the font type, check the active states and setup the desktop view.