Update
I was able to add the fontAwesome icon kit to this project by following the documentation, adding it into the script tag adn calling the icons in an unordered  list. From there, I was able to outline the icons and border-radius them so that they show up as a circular border. I used flexbox to make them inline, centered and spaced out somewhat. I also set the main container to 100dvw and 100dvh so the size on mobile looks better. Will do more to adjust mobile view before working on the desktop.